refactor(linebreak): give the compiler DOM-free seams
compileBlock read two DOM-typed fields: `run.wrappers` through codeWrapper,
and `block.wrappers` through runEdgeWidths. Both become optional callbacks on
CompileContext — `isCode` and `edgesFor` — with linebreaker.ts passing the
same two functions, so the DOM path computes exactly what it computed before.

The four pure helpers the compiler shared with the extractor move to
src/text/source.ts: hasVisibleText, collapseWhitespace, breakAllowedAt and
SourceRange. None of them touches a node.

compile.ts's runtime import graph is now 11 files, zero of them under dom/ and
zero bare packages, down from 17 files reaching dom/extract; cold import is
3.69 ms median over 9 fresh processes. The three snapshot digests are
byte-identical, the 8 parity gates stay green, and Chromium e2e is 27/27.

Callers that build an ExtractedBlock themselves now wire the two callbacks;
the test support module does, mirroring linebreaker.ts.
